[VarExporter] Leverage deepclone_hydrate() in Hydrator and Instantiator
