--- src/Symfony/Component/VarExporter/Internal/Values.php
@@ -1,25 +1,0 @@
-<?php
-
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Symfony\Component\VarExporter\Internal;
-
-/**
- * @author Nicolas Grekas <p@tchwork.com>
- *
- * @internal
- */
-class Values
-{
-    public function __construct(
-        public readonly array $values,
-    ) {
-    }
-}
