+++ src/Symfony/Component/VarExporter/Tests/ProxyHelperTest.php
@@ -150,7 +150,6 @@ class ProxyHelperTest extends TestCase
             }
 
             // Help opcache.preload discover always-needed symbols
-            class_exists(\Symfony\Component\VarExporter\Internal\Hydrator::class);
             class_exists(\Symfony\Component\VarExporter\Internal\LazyObjectRegistry::class);
 
             EOPHP;
@@ -196,7 +195,6 @@ class ProxyHelperTest extends TestCase
             }
 
             // Help opcache.preload discover always-needed symbols
-            class_exists(\Symfony\Component\VarExporter\Internal\Hydrator::class);
             class_exists(\Symfony\Component\VarExporter\Internal\LazyObjectRegistry::class);
 
             EOPHP;
